Get tokens from the environment when commands are invoked
Instead of on import of the top level module. This makes testing
easier and allows modules to be imported in any order you want.

This PR also uses a pytest fixture to set tokens in the environment
for individual tests. They are unset at the end, so this should
prevent a few categories of future testing bugs.